Added a feature to read bot names from a text file instead of a hardcoded list in the code.
Previously, the names of the bots were hardcoded in the list called BOT_NAMES. To make it easier to change the names of the bots, a new feature has been added to read the names from a text file called "bots.txt". This file contains a list of names, with one name per line. The code now reads this file and creates bot players with the names from the file. This makes it much easier to update the list of bot names without having to modify the code.