Added implementation for two hands and a message to inform intermediate and advanced players about the process of using two hands when the player gets a card with a value of 10 on the first hand.
In this commit, we added the implementation of the two-hand feature in our blackjack game. However, the second hand is still in development. We also added a message to inform intermediate and advanced players that they can use two hands when they get a card with a value of 10 on their first hand. This is just the beginning of the two-hand process, and we will continue to develop this feature in future commits.

Additionally, we updated the code to only allow intermediate and advanced players to use the 10-value card feature, as it requires more advanced strategy. We also added logging statements to inform the player of their chances of getting a blackjack, and to show the dealer's first card.

These changes will improve the gameplay experience for intermediate and advanced players, and pave the way for future updates to the two-hand feature.