Fixing "strip error" issue
Previously, there was an issue in the code where the game would move to the second hand instead of splitting the hand when the player drew their second card. This issue has now been fixed.

To avoid confusion, the code has also been updated to use English variable and function names. Additionally, a global variable named BLACKJACK has been added to the code to represent the value of 21 in Blackjack, and the name of this variable has been improved to better reflect its purpose.